Use aliases for fields and call collectElts for bundles in Wire and Reg
This fixes the issue with not being able to generate proper field access
for temporary bundles (bundles not created at the top-level of a
module). The solution is to have the imm value in Field point to aliases
instead of direct refs and call setRefs on bundles when they are passed
into Reg or Wire.